Allow Hashes to be split
This is a proposed change to allow Hashes to be split.  An event is created for each key and the @target or @field on the new event is set to a hash containing the key => value pair.  This causes one large hash to be split into many smaller hashes, one event per key.